+++ tests/methcla_tests.hpp
@@ -100,13 +100,6 @@ namespace Methcla { namespace Tests {
 
     inline static void sleepFor(double seconds)
     {
-#if defined(__native_client__)
-        struct timespec ts;
-        ts.tv_sec = seconds;
-        ts.tv_nsec = (seconds - ts.tv_sec) * 1e9;
-        nanosleep(&ts, &ts);
-#else
         std::this_thread::sleep_for(std::chrono::duration<double>(seconds));
-#endif
     }
 }} // namespace Methcla::Tests